Move payload conversion code to separate module
